增删改查并做后端数据校验

+++ express-mysql/routes/users.js
@@ -26,8 +26,12 @@ router.get('/changeUser', function (req, res, next) {
         if (err) {
             res.send(BizResult.fail(err))
         } else {
-            // 将 MySQL 查询结果作为路由返回值
-            res.send(BizResult.success(result))
+            if (result.affectedRows === 0){
+                console.log('result',result)
+                res.send(BizResult.fail('请输入正确的id'))
+            }else{
+                res.send(BizResult.success('修改成功'))
+            }
         }
     })
 });
@@ -35,6 +39,28 @@ router.get('/changeUser', function (req, res, next) {
 router.post('/addUser', function (req, res, next) {
     console.log('req.body', req.body)
     let data = req.body;
+
+    for (let item in data){
+        console.log('item',item)
+        if (!data[item]){
+            res.send(BizResult.fail('请输入完整信息'))
+            return;
+        }
+    }
+
+    connection?.query(`select count(*) from user where id=${data.id}`, (err, result) => {
+        if (err) {
+            res.send(BizResult.fail(err))
+        } else {
+            console.log('查询id是否存在',result)
+            if (result[0]['count(*)'] > 0) {
+                res.send(BizResult.fail('用户已存在'));
+                return;
+            }
+        }
+    })
+
+
     const sql = `insert into user value(${data.id},'${data.name}','${data.phone}','${data.keywords}')`
     /* 使用 connection.query 来执行 sql 语句 */
     // 第一个参数为 sql 语句，可以透过 js 自由组合
@@ -44,7 +70,7 @@ router.post('/addUser', function (req, res, next) {
             res.send(BizResult.fail(err))
         } else {
             // 将 MySQL 查询结果作为路由返回值
-            res.send(BizResult.success(result))
+            res.send(BizResult.success('新增成功'))
         }
     })
 });
@@ -52,7 +78,14 @@ router.post('/addUser', function (req, res, next) {
 router.delete('/deleteUser', function (req, res, next) {
     console.log('req.body', req.body)
     let data = req.body;
+
+    if (!data.id){
+        res.send(BizResult.fail('id 不能为空'));
+        return;
+    }
+
     const sql = `delete from user where id=${data.id}`
+
     /* 使用 connection.query 来执行 sql 语句 */
     // 第一个参数为 sql 语句，可以透过 js 自由组合
     // 第二个参数为回调函数，err 表示查询异常、第二个参数则为查询结果（这里的查询结果为多个用户行）
@@ -60,8 +93,14 @@ router.delete('/deleteUser', function (req, res, next) {
         if (err) {
             res.send(BizResult.fail(err))
         } else {
+            //insertId 是插入的 ID；affectedRows 是受影响的行数；changedRows 返回更新了多少行
+            if (result.affectedRows === 0){
+                console.log('result',result)
+                res.send(BizResult.fail('该用户不存在'))
+            }else{
+                res.send(BizResult.success('删除成功'))
+            }
             // 将 MySQL 查询结果作为路由返回值
-            res.send(BizResult.success(result))
         }
     })
 });
